--- dom_sink/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "html5ever_dom_sink"
-version = "0.2.0"
-authors = [ "The html5ever Project Developers" ]
-description = "Basic DOM implementation for html5ever"
-license = "MIT / Apache-2.0"
-repository = "https://github.com/servo/html5ever"
-
-[lib]
-name = "html5ever_dom_sink"
-
-[dependencies]
-mac = "0"
-string_cache = "0"
-tendril = "0.1"
-
-[dev-dependencies]
-string_cache_plugin = "0"
-
-[dependencies.html5ever]
-version = "0.2"
-path = "../"
-
-[dev-dependencies.html5ever_test_util]
-version = "0.1.0"
-path = "../test_util"
