+++ src/lib.rs
@@ -10,11 +10,13 @@
 #![crate_name="html5ever"]
 #![crate_type="dylib"]
 
-#![cfg_attr(feature = "unstable", feature(plugin))]
+#![cfg_attr(feature = "unstable", feature(plugin, rc_weak))]
 #![cfg_attr(test, deny(warnings))]
 #![allow(unused_parens)]
 
 #![cfg_attr(feature = "unstable", plugin(string_cache_plugin))]
+
+extern crate rc;
 
 #[macro_use]
 extern crate log;
@@ -47,7 +49,6 @@ mod util {
 
 pub mod tokenizer;
 pub mod tree_builder;
-
 pub mod serialize;
-
 pub mod driver;
+pub mod rcdom;
